pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "content-stack"
version = "0.0.1"
description = "Globally-installed Python daemon (FastAPI + SQLite/WAL + MCP Streamable HTTP) plus Vue UI giving any LLM client a stateful CRUD seam over multi-project SEO content pipelines."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "content-stack contributors" }]
keywords = ["mcp", "fastapi", "sqlite", "seo", "content", "daemon"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "mcp>=1.0",
    "httpx>=0.27",
    "cryptography>=43",
    "apscheduler>=3.10",
    "pydantic-settings>=2.6",
    "structlog>=24.4",
    "typer>=0.13",
    "python-dotenv>=1.0",
]

[project.scripts]
content-stack = "content_stack.cli:app"

[project.urls]
Homepage = "https://github.com/flowmonkey-io/content-stack"
Repository = "https://github.com/flowmonkey-io/content-stack"

[dependency-groups]
dev = [
    "ruff>=0.8",
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.34",
    "mypy>=1.13",
    "freezegun>=1.5",
]

[tool.hatch.build.targets.wheel]
packages = ["content_stack"]

[tool.hatch.build.targets.wheel.force-include]
"content_stack/ui_dist" = "content_stack/ui_dist"

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
    "content_stack/db/migrations/versions",
    "content_stack/ui_dist",
    ".upstream",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
# B008: ``= Depends(...)`` defaults are the canonical FastAPI dep pattern.
# A002: ``status`` shadows a stdlib name but matches the column it filters on,
#       and we accept it as the natural query-param spelling.
ignore = ["B008", "A002"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]
"content_stack/api/**" = ["A002"]

[tool.mypy]
python_version = "3.12"
strict = false
# SQLModel exposes column descriptors at class-attribute level; mypy
# sees the runtime value type (``bool``, ``datetime``, etc.) and flags
# the SA-level methods (``.is_()``, ``.asc()``, ``.in_()``, ``.desc()``).
# Repository code uses these freely under per-line ignores, but the
# narrow code matrix mypy reports per call doesn't always agree across
# minor versions. We therefore allow unused ignores rather than burn
# code review cycles tracking the exact ignore code per line.
warn_unused_ignores = false
no_implicit_optional = true
ignore_missing_imports = true
exclude = [
    "content_stack/db/migrations/",
    "content_stack/ui_dist/",
    "tests/",
    "build/",
    "dist/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-ra", "--strict-markers"]
markers = [
    "benchmark: marks acceptance benchmarks (audit B-07 100x set_draft, etc.)",
]
# `schema_json` is the canonical PLAN.md L361 column name on the
# `schema_emits` table; SQLModel's parent class exposes a `schema_json`
# attribute that triggers this benign UserWarning at class-construction time.
filterwarnings = [
    "ignore:Field name \"schema_json\" in \"SchemaEmit\" shadows an attribute in parent \"SQLModel\":UserWarning",
    "ignore:Field name \"schema_json\" in \"SchemaEmitOut\" shadows an attribute in parent \"BaseModel\":UserWarning",
    "ignore:Field name \"schema_json\" in \"SchemaSetRequest\" shadows an attribute in parent \"BaseModel\":UserWarning",
    "ignore:Field name \"schema_json\" in \"SchemaSetInput\" shadows an attribute in parent \"MCPInput\":UserWarning",
]
